backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
email-validator==2.1.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
